requirements.txt
aiocontextvars==0.2.2
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
colorlog==4.0.2
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
DingtalkChatbot==1.3.0
Django==3.2
django-celery==3.3.1
django-cors-headers==3.7.0
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-rest-swagger==2.2.0
django-simpleui==2020.7
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
har2case==0.3.1
HttpRunner==1.5.15
idna==2.8
immutables==0.14
inflection==0.5.0
itypes==1.1.0
Jinja2==2.10.3
kombu==3.0.37
loguru==0.5.1
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.10.3
pytz==2019.3
PyYAML~=5.3.1
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
tornado==4.5.3
uritemplate==3.0.1
urllib3==1.25.7
#uWSGI==2.0.18
django_mysql
json5==0.9.5
crontab~=0.22.9
django_bulk_update==2.2.0
#uWSGI==2.0.18
PyMySQL==0.10.1
jsonpath==0.82
Faker==4.9.0
pytest==6.1.2
DBUtils==2.0
rsa==4.6
pytesseract==0.3.6
openpyxl==3.0.5
logger==1.4
allure-pytest==2.8.19
allure-python-commons==2.8.19
cv2-utils==0.2.0
opencv_python==4.5.2.52
numpy==1.20.3
xlrd==1.2.0
xlwt==1.3.0
#dwebsocket==0.4.2
dwebsocket==0.5.12
python-ldap==3.3.1
django-auth-ldap==2.4.0
pyldap==2.4.45
gunicorn==20.1.0
paramiko==2.7.2
channels==3.0.3
uvicorn==0.14.0
werkzeug==2.0.1
python-nacos==0.1.1
concurrent_log_handler==0.9.19
#salt==3003.1
pandas==1.2.4
redis==2.10.6
django-influxdb-metrics
xmltodict
jsonpath_rw
python-jenkins==1.7.0

